Add files via upload
In this mini project, you are required to create an image gallery using React. The gallery should fetch a set of
images from given api, limit the results to 20 images, and display them in a user-friendly UI. Upon clicking on
any photo, the user should be redirected to a single image view, where additional information, such as the title
and description of the image, is shown.